U-DRS/requirements.txt
# U-DRS: Universal Image-to-3D Damage Reconstruction System
# Requirements

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0
imageio>=2.31.0

# 3D Processing
open3d>=0.17.0
trimesh>=3.23.0
PyMCubes>=0.1.4

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.16.0
seaborn>=0.12.0

# API Framework
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
aiofiles>=23.0.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Model Optimization (Optional)
# onnx>=1.14.0
# onnxruntime>=1.15.0
# tensorrt>=8.6.0  # Requires separate installation

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
